Switchboard component plus individual component-related channel bundles
These individual channel bundle components are somewhat like the interface
for individual components, in the way that an engine's wiring harness
represents the interface of an engine. The switchboard then only needs to know
the channel bundles, without needing any knowledge about a specific implementation.
Despite being a little longer, this is cleaner and less tightly coupled than the
previous version were there was one centralized channels component, of which the
other components needed to know the names of individual channels.