introduction of interop component for Redis Pub/Sub
The interop component allows sending and receiving messages via
Redis Pub/Sub. It has both a :send and a :receive channel and can
be used on both sides of the Pub/Sub.

In this commit, the application is still complete but all
messages are sent over Pub/Sub so the next step will be to
separate this into two applications.