Persistence of updated RTs removed
Persisting changes to RTs (like RT and favorite counts) over and over again
results in the tweets in ElasticSearch changing all the time, which constantly
invalidates caches. Also in the case of a network partition in the ES cluster,
bets are much better that the system will remain consistent if new items are only
ever inserted but never updated. See the Jepsen post on ElasticSearch.

http://aphyr.com/posts/317-call-me-maybe-elasticsearch

The actual changes can always be reconstructed from the retweet status of the
retweets, which always hold the latest version at the time.