fixing issue with start / stop of Switchboard component
Before this fix, restarting the entire systems on the REPL didn't
work because apparently, the component library did not detect changes
in the Switchboard component when no key/value pairs where assoc'd
on the component map on startup, so it wasn't actually restarted.
That, however, left the old channels wired instead of the new channels.