Dependence on state namespace removed in ui.tweets and ui.elements namespaces
Instead of directly accessing the state, the ui components receives updates
via a channel connected to a mult that broadcasts state changes.

Also removed: dependence of the state namespace on the Reagent atom.

In addition to protecting state by not modifying it outside the state namespace any
longer, this also allows only updating tweets when the "Live" switch is on. When it's
off, the users can browse through tweet lists without being disturbed by fast updates.