Protect application state from access and alteration
By keeping the application state atom inside a let binding, it cannot be accessed
from the outside, thus protecting it from (accidental) mutilation from the outside.
Instead, the only way to interact with the application state shall be through
putting messages put on the channels passed as parameters in the init-state function.